# app/public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "shop_car",
        "description": "shop car documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api/shop/v1/",
            "description": "test"
        }
    ],
    "paths": {
        "/cars/list": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get filtered list of cars",
                "description": "return list of cars",
                "parameters": [
                    {
                        "name": "city_id",
                        "in": "query",
                        "description": "city's id of car",
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "car_model_id",
                        "in": "query",
                        "description": "model's id of car",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "price of car. Between min and max price",
                        "schema": {
                            "type": "json",
                            "example": "[1000000,100000000]"
                        }
                    },
                    {
                        "name": "type_drive",
                        "in": "query",
                        "description": "type drive of car",
                        "schema": {
                            "type": "string",
                            "example": "full_drive"
                        }
                    },
                    {
                        "name": "type_engine",
                        "in": "query",
                        "description": "type engine of car",
                        "schema": {
                            "type": "string",
                            "example": "petrol_engine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/cars/detail": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get detail info of car",
                "description": "return detail info car",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id of car",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/cars/brands/list": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get list brands",
                "description": "return list brands",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/cars/models/list": {
            "get": {
                "tags": [
                    "Models"
                ],
                "summary": "Get filtred list models",
                "description": "return list brands",
                "parameters": [
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "id of brand",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/cars/create": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "create new car",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarDtoObject"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "created"
                    }
                }
            }
        },
        "/cars/delete": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "delete car",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DetailCarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "deleted"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CarDtoObject": {
                "title": "CarDtoObject",
                "description": "car-dto object",
                "required": [
                    ""
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of car",
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "title": "description",
                        "description": "description of car",
                        "type": "string",
                        "example": "The best car in the world"
                    },
                    "preview": {
                        "title": "preview",
                        "description": "preview of car",
                        "type": "file"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of car",
                        "type": "integer",
                        "example": 4000000
                    },
                    "creator_id": {
                        "title": "creator_id",
                        "description": "creator's id of car",
                        "type": "integer",
                        "example": 1
                    },
                    "car_model_id": {
                        "title": "car_model_id",
                        "description": "model's id of car",
                        "type": "integer",
                        "example": 1
                    },
                    "city_id": {
                        "title": "city_id",
                        "description": "city's id of car",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "FilterCarDtoObject": {
                "title": "FilterCarDtoObject",
                "description": "filter-car-dto object",
                "required": [
                    ""
                ],
                "properties": {
                    "city_id": {
                        "title": "city_id",
                        "description": "city id",
                        "type": "integer",
                        "example": 1
                    },
                    "type_drive": {
                        "title": "type_drive",
                        "description": "type drive of car",
                        "type": "string",
                        "example": "full_drive"
                    },
                    "type_engine": {
                        "title": "type_engine",
                        "description": "type engine of car",
                        "type": "string",
                        "example": "petrol_engine"
                    },
                    "car_model_id": {
                        "title": "car_model_id",
                        "description": "model's id of car",
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "title": "price",
                        "description": "price of car. Between min and max price",
                        "type": "json",
                        "example": "[1000000,100000000]"
                    }
                },
                "type": "object"
            },
            "DetailCarRequest": {
                "title": "DetailCarRequest",
                "description": "car-dto object",
                "required": [
                    ""
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of car",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    }
}